+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [TBD]
+
+### Added
+- Group by type setting to organize search results into Tabs and Bookmarks sections (on by default)
+
 ## [1.3.0] - 2026-02-05
 
 ### Changed